requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.8.0
python-docx>=1.1.0

